APIRoute class to strip apikey param FOL-140
Custom APIRoute class [1] strips the trailing `&apikey=` hardcoded in
SpineOMatic request [2] before passing on the request to FOLIO endpoint.

This is necessary to pass validation since the barcode in the path must
be an integer.

Using Python's String.partition() to split the barcode at `&`, so not
targeting the full `&apikey=` appended string which hopefully will
result in reduced maintenance if/when changes are introduced in
SpineOMatic.

[1]: https://fastapi.tiangolo.com/advanced/custom-request-and-route/#custom-apiroute-class-in-a-router
[2]: https://github.com/ExLibrisGroup/SpineOMatic/blob/f31d0f2b44356d9a13524337de45d108f82735e7/Form1.vb#L1218