Return an object as the root node
It appears that SpineOMatic is expecting an object at the root node.
Since FOLIO's `/inventory/items` endpoint always returns a list [1],
we need to trim it down to the item itself before returning a
tailored response suitable for consumption by SpineOMatic.

Making use of xml2json custom wrapper [2] to set the root node.

[1] https://s3.amazonaws.com/foliodocs/api/mod-inventory/inventory.html#inventory_items_get
[2] https://json2xml.readthedocs.io/en/latest/#custom-wrappers-and-indent