Simplify holdings, no perm or temp call numbers
Rely on FOLIO's derived/computed `effectiveCallNumberComponents`, which
will be populated for ALL items, and map these components to their
corresponding nodes in Alma's rest_item.xsd [1].

EBSCO had an issue with one of our early data loads which resulted in
many (most?) items missing `effectiveCallNumberComponents`. This has
since been resolved, but only reflected on `cornell-test` at this point,
so we'll be pointing the middleware to `cornell-test` instead of
`cornell-training` as part of this deploy as well.

[1]: https://developers.exlibrisgroup.com/alma/apis/docs/xsd/rest_item.xsd/?tags=GET#holding_data