Transform xml: align with Alma's response FOL-142 (#4)
[skip ci] – feature branch has already been deployed. [1]

Introduce xslt based on example response from Alma's RESTful API [1] to
transform the straight-up FOLIO json -> xml conversion to one that can
be validated by Alma's rest_item.xsd schema [2].

As originally reported in #3, this is to ensure that the XML response
from the middleware can be consumed by SpineOMatic, which currently is
dependent on receiving XML conforming to the Alma's item schema, so that
it can then transform it to the legacy SOAP format [3] before it can
finally be used.

We thought we could map to any XML using SpineOMatic's custom formatting
capability, but it turns out that checking the "Don't convert XML
option" suspends the app from normal operation [4].

[1]: https://api-na.hosted.exlibrisgroup.com/almaws/v1/items?item_barcode=75912&view=label&apikey=l7xx2af7939c63424511946e0fcdc35fe22a
[2]: https://developers.exlibrisgroup.com/alma/apis/docs/xsd/rest_item.xsd/?tags=GET
[3]: https://github.com/ExLibrisGroup/SpineOMatic/blob/f31d0f2b44356d9a13524337de45d108f82735e7/Form1.vb#L1322-L1368
[4]: https://github.com/ExLibrisGroup/SpineOMatic/blob/f31d0f2b44356d9a13524337de45d108f82735e7/Form1.vb#L1247-L1298

* New transform query param

Defaults to True but can be set to False to bypass the XSLT transform
(to Alma's item schema) and return the raw XML that is generated after
converting the FOLIO json.

* Use 0 to indicate LOC for call_number_type

Temporary workaround.

* Use desc attribute to specify shelving scheme

This is critical because SpineOMatic's internal transform of the XML
splits:
```
<call_number_type desc="Library of Congress classification">0</call_number_type>
```

into the following two nodes:
```
<call_number_type_name>Library of Congress classification</call_number_type_name>
<call_number_type>0</call_number_type>
```

For the attribute value, we're following CallNumberType code-table
defined in Alma's rest_item.xsd [1] which are based on the LOC MARC 21
spec for the `852` [2].

Again, we're currently hardcoding to LOC, since this covers a majority
of our items, but we'll come back and actually map the FOLIO
effectiveCallNumberComponents.typeID UUID.

[1]: https://developers.exlibrisgroup.com/alma/apis/docs/xsd/rest_item.xsd/?tags=GET#holding_data
[2]: https://www.loc.gov/marc/holdings/hd852.html

* Simplify holdings, no perm or temp call numbers

Rely on FOLIO's derived/computed `effectiveCallNumberComponents`, which
will be populated for ALL items, and map these components to their
corresponding nodes in Alma's rest_item.xsd [1].

EBSCO had an issue with one of our early data loads which resulted in
many (most?) items missing `effectiveCallNumberComponents`. This has
since been resolved, but only reflected on `cornell-test` at this point,
so we'll be pointing the middleware to `cornell-test` instead of
`cornell-training` as part of this deploy as well.

[1]: https://developers.exlibrisgroup.com/alma/apis/docs/xsd/rest_item.xsd/?tags=GET#holding_data

* Use desc attribute for library and location nodes

This is critical because SpineOMatic's internal transform of the XML
expects the desc attribute for both of these nodes and uses it to create
a new node in the transform output

For example, this <library> node:
```
<library desc="Olin Library">olin</library>
```

is split into the following two nodes:
```
<library_code>olin</library_name>
<library_name>Olin Library</library_code>
```

The same pattern applies for the `<location> node.

For now, we'll be repeating the same value
(`item.effectiveLocation.name`) for the desc attribute and the node
itself which should satisfy our needs, according to Peter M.

* Dynamic desc attribute for physical_material_type

Sinc SpineOMatic internally transforms this node as well, it's important
we're not hardcoded to book and can support serial as well.

* Dynamic desc attribute for base_status

SpineOMatic also internally transforms and splits out this node.
Hardcoding the node value to 1 for now, since we don't have an
equivalent in FOLIO and this shouldn't have any repercussions on spine
labeling.

* Raw & transformed XML from Alma sample endpoint

Used as reference to reverse engineer internal SpineOMatic
transformation.

* Drop `parsed_call_number` from the xslt

We're not using it and we suspect the empty child nodes (with self
closing tags) may be behind the following exception [1]:

```
SpineOMatic cannot process the XML due to invalid characters in the data.
The error returned is:
System.Xml.XmlException: '_1' is an unexpected token. The expected token is '>'...
    ...
    at SpineLabeler.Form1.getBarcodeFile()
```

[1]: https://github.com/ExLibrisGroup/SpineOMatic/blob/5aafb40bd4268259242d7eb279496f95eee4ece3/Form1.vb#L1458-L1490

* Actions: Temp switch trigger to this branch

* Revert back to trigger on push to main

[1]: https://github.blog/changelog/2021-02-08-github-actions-skip-pull-request-and-push-workflows-with-skip-ci/